gomaasapi: use a giant honking lock around all API calls
The test server mock has a lot of mutable state, none of it guarded
by locks. It is possible to go through each field and check who uses
it, find all their methods, add locks, rinse, repeat.

Or, we could do something really dirty and serialise all API access
through one mutex. This has the obvious side effect that API calls
are no longer concurrent, but was their any promise of that before?
The change also has the effect that a memory barrier, covering all
the state in the test server structure, is asserted before and after
each API call