Constraint matches are lists
In APIv2 the constraint matches are actually label -> [id, id...]. This
makes sense because in v1 they were id -> label, which meant that label
could appear more than once; there could be multiple interfaces or
storage devices that fulfil a specific constraint.

Change ConstraintMatches and the parsing to handle this.