Optional HardwareInfo
Unfortunately, HardwareInfo isn't always available for every MAAS
version. The prior PR[1] expected that hardware_info was available. The
fix is simple, just make it optional when attempting to validate it with
the schema coercing.

Additionally, I've added a few more tests to ensure that we're testing
various scenarios of hardware_info being omitted. Lastly, it felt like
there was a missing check for TestReadMachines for validating the
hardware info for a given machine response.

This should fix[2] in Juju.

 1. https://github.com/juju/gomaasapi/pull/96
 2. https://bugs.launchpad.net/juju/+bug/2009064